client just about done, except...
we need to fix the capnproto backend for serializing replica messages;
unbeknownst to me, it only works in !Send environments, which severely
limits its usage in tokio spawned tasks, unless they're marked with a
runtime specific invariant (based around a LocalSet type) that enables a
set of futures to be run on the same thread through its execution cycle.

the natural alternative is to skip using capnproto, but it claims to
score better than any other serialization library in a few benchmarks
I've consulted, which draws me towards using it, despite having to
invest time in finding a solution around the !Send limitation.

if I don't find a solution in a useful time frame of a day or so, I'll
skip capnproto altogether and move to the serde framework, with msgpack
as the serialization backend, which shouldn't be too hard, since I've
already coded an abstraction for serializing replica messages.