fix bootstrap code
because of the nature of how the network code is shared between client
and replica nodes in febft, a behavior which was deemed correct for
replicas unfortunately posed an error for client nodes. channels from
the request batcher code were dropped in clients, which caused annoying
runtime errors when using futures channels. curiously, with mpmc channel
implementations this was not an issue, hence the switch to flume.
regardless, the bug was fixed in this commit, so any of the currently
supported channel implementations work fine in febft.