fix consensus code
In this branch, we hunted for bugs in the consensus code. The first one
resulted from the fact we were not queueing some of the received client
requests in a new consensus instance, which resulted in stale clients.
Secondly, we had faulty logic in the verification of speculative commits
generated ahead of time.